+++ Data/PvPEvents.lua
@@ -12,8 +12,8 @@ local function bracket(level)
         level = level,
         contact = contact,
         events = level == 19 and {
-            { battleground="WARSONG GULCH", realm="FIREMAW CLUSTER", scope="ALL CONNECTED REALMS", date="15 AUG 2026", time="16:00-21:00 SERVER TIME", startsAt=1786802400, endsAt=1786820400, faction="HORDE VS ALLIANCE" },
-            { battleground="WARSONG GULCH", realm="FIREMAW CLUSTER", scope="ALL CONNECTED REALMS", date="16 AUG 2026", time="16:00-21:00 SERVER TIME", startsAt=1786888800, endsAt=1786906800, faction="HORDE VS ALLIANCE" },
+            { battleground="WARSONG GULCH", queueLabel="QUEUE FOR WSG", realm="FIREMAW CLUSTER", scope="ALL CONNECTED REALMS", date="15 AUG 2026", time="16:00-21:00 SERVER TIME", startsAt=1786802400, endsAt=1786820400, faction="HORDE VS ALLIANCE" },
+            { battleground="WARSONG GULCH", queueLabel="QUEUE FOR WSG", realm="FIREMAW CLUSTER", scope="ALL CONNECTED REALMS", date="16 AUG 2026", time="16:00-21:00 SERVER TIME", startsAt=1786888800, endsAt=1786906800, faction="HORDE VS ALLIANCE" },
         } or {},
     }
 end